Move OperationMode to Operation
