+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/InteractionModel.kt
@@ -5,8 +5,8 @@ import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.spring
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.Density
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode.KEYFRAME
+import com.bumble.appyx.interactions.core.Operation.Mode
+import com.bumble.appyx.interactions.core.Operation.Mode.KEYFRAME
 import com.bumble.appyx.interactions.core.inputsource.AnimatedInputSource
 import com.bumble.appyx.interactions.core.inputsource.DebugProgressInputSource
 import com.bumble.appyx.interactions.core.inputsource.DragProgressInputSource
@@ -81,7 +81,7 @@ open class InteractionModel<NavTarget : Any, ModelState : Any>(
     fun operation(
         operation: Operation<ModelState>,
         animationSpec: AnimationSpec<Float> = defaultAnimationSpec,
-        mode: OperationMode = KEYFRAME,
+        mode: Mode = KEYFRAME,
     ) {
         when {
             isDebug -> debug.operation(operation)
