+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/Operation.kt
@@ -5,7 +5,12 @@ import com.bumble.appyx.interactions.Parcelize
 
 interface Operation<ModelState> : (ModelState) -> NavTransition<ModelState>, Parcelable {
 
+    enum class Mode {
+        KEYFRAME, IMMEDIATE
+    }
+
     fun isApplicable(state: ModelState): Boolean
+
 
     @Parcelize
     class Noop<ModelState> : Operation<ModelState> {
@@ -17,7 +22,6 @@ interface Operation<ModelState> : (ModelState) -> NavTransition<ModelState>, Par
             NavTransition(state, state)
 
         override fun equals(other: Any?): Boolean = this.javaClass == other?.javaClass
-
         override fun hashCode(): Int = this.javaClass.hashCode()
     }
 }
