+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/inputsource/AnimatedInputSource.kt
@@ -6,9 +6,9 @@ import androidx.compose.animation.core.spring
 import com.bumble.appyx.interactions.Logger
 import com.bumble.appyx.interactions.core.Operation
 import com.bumble.appyx.interactions.core.TransitionModel
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode.KEYFRAME
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode.IMMEDIATE
+import com.bumble.appyx.interactions.core.Operation.Mode
+import com.bumble.appyx.interactions.core.Operation.Mode.KEYFRAME
+import com.bumble.appyx.interactions.core.Operation.Mode.IMMEDIATE
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -24,13 +24,13 @@ class AnimatedInputSource<NavTarget : Any, ModelState>(
     private val animatable = Animatable(0f)
     // FIXME private lateinit var result: AnimationResult<Float, AnimationVector1D>
 
-    override fun operation(operation: Operation<ModelState>, mode: OperationMode) {
+    override fun operation(operation: Operation<ModelState>, mode: Mode) {
         operation(operation, mode, defaultAnimationSpec)
     }
 
     fun operation(
         operation: Operation<ModelState>,
-        mode: OperationMode = KEYFRAME,
+        mode: Mode = KEYFRAME,
         animationSpec: AnimationSpec<Float>
     ) {
         Logger.log("AnimatedInputSource", "New operation: $operation")
