+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/inputsource/DebugProgressInputSource.kt
@@ -19,7 +19,7 @@ class DebugProgressInputSource<NavTarget, ModelState>(
     private lateinit var result: AnimationResult<Float, AnimationVector1D>
     private var progress: Float = 1f
 
-    override fun operation(operation: Operation<ModelState>, mode: TransitionModel.OperationMode) {
+    override fun operation(operation: Operation<ModelState>, mode: Operation.Mode) {
         // Regardless of mode, only enqueue makes sense – consider if API could better reflect that
         navModel.enqueue(operation)
     }
