+++ appyx-interactions/common/src/commonMain/kotlin/com/bumble/appyx/interactions/core/inputsource/InputSource.kt
@@ -1,10 +1,10 @@
 package com.bumble.appyx.interactions.core.inputsource
 
 import com.bumble.appyx.interactions.core.Operation
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode
-import com.bumble.appyx.interactions.core.TransitionModel.OperationMode.KEYFRAME
+import com.bumble.appyx.interactions.core.Operation.Mode
+import com.bumble.appyx.interactions.core.Operation.Mode.KEYFRAME
 
 interface InputSource<NavTarget, ModelState> {
 
-    fun operation(operation: Operation<ModelState>, mode: OperationMode = KEYFRAME)
+    fun operation(operation: Operation<ModelState>, mode: Mode = KEYFRAME)
 }
